+++ test/lsu/test_pma.py
@@ -1,9 +1,9 @@
 from amaranth.sim import Settle
-from coreblocks.func_blocks.lsu.pma import PMAChecker, PMARegion
+from coreblocks.func_blocks.fu.lsu.pma import PMAChecker, PMARegion
 
 from transactron.lib import Adapter
 from coreblocks.params import GenParams
-from coreblocks.func_blocks.lsu.dummyLsu import LSUDummy
+from coreblocks.func_blocks.fu.lsu.dummyLsu import LSUDummy
 from coreblocks.params.configurations import test_core_config
 from coreblocks.frontend.decoder import *
 from coreblocks.interface.keys import ExceptionReportKey
